rollback some of the splitting in the hope to get allocations back under control but to no avail
