+++ benchrunner/Benchrunner.hs
@@ -5,8 +5,12 @@ module Main where
 import           Data.List.Split    ( splitOn )
 import           System.Environment ( getArgs )
 import           Control.Monad      ( unless )
+import           Control.DeepSeq    ( NFData )
 
 import qualified Array as A
+import           Linear.Common
+
+import           ForeignFunctionImports as FFI
 import qualified Vector as V
 import           Input
 import qualified Measure as M
@@ -14,6 +18,58 @@ import           Sort
 import           Utils
 import qualified Microbench as MB
 
+import qualified Insertion as I
+import qualified QuickSort as Q
+import qualified DpsMergeSort4 as DMS
+import qualified DpsMergeSort4Par as DMSP
+import qualified PiecewiseFallbackSort as PFS
+import qualified PiecewiseFallbackSortPar as PFSP
+import qualified Data.Vector.Algorithms.Insertion as ISDVS
+import qualified Data.Vector.Algorithms.Merge as MSDVS
+import qualified Data.Vector.Algorithms.Intro as QSDVS
+
+--
+-- Table of sorting functions
+--
+
+sortFn :: (Show a, A.HasPrimOrd a, NFData a) => SortAlgo -> ParOrSeq -> (A.Array a -. A.Array a)
+sortFn bench parorseq = case (bench,parorseq) of
+  (Insertionsort, Seq) -> I.isort_top'
+  (Quicksort, Seq)     -> Q.quickSort'
+  (Mergesort, Seq) -> DMS.msort
+  (Mergesort, Par) -> DMSP.msort
+  (Optsort,   Seq) -> PFS.pfsort
+  (Optsort,   Par) -> PFSP.pfsort
+  oth -> error $ "sortFn: unknown configuration: " ++ show oth
+{-# INLINABLE sortFn #-}
+
+vectorSortFn :: SortAlgo -> ParOrSeq -> VecSort
+vectorSortFn bench parorseq = case (bench,parorseq) of
+  (Insertionsort, Seq) -> ISDVS.sort
+  (Mergesort,     Seq) -> MSDVS.sort
+  (Quicksort,     Seq) -> QSDVS.sort
+  oth -> error $ "sortFn: unknown configuration: " ++ show oth
+{-# INLINABLE vectorSortFn #-}
+
+sortFnC :: SortAlgo -> FFI.SortFn
+sortFnC alg = case alg of
+                    Insertionsort -> FFI.c_insertionsort
+                    Mergesort -> FFI.c_mergesort
+                    Quicksort -> FFI.c_quicksort
+                    _ -> error "sortFnC: Csort not implemented!"
+{-# INLINABLE sortFnC #-}
+
+sortFnCxx :: SortAlgo -> FFI.SortFnCxx
+sortFnCxx alg = case alg of
+                    Insertionsort -> FFI.cxx_int_insertionsort
+                    Mergesort -> FFI.cxx_int_mergesort
+                    Quicksort -> FFI.cxx_int_quicksort
+                    _ -> error "sortFnCxx: Csort not implemented!"
+{-# INLINABLE sortFnCxx #-}
+
+--
+-- Select which benchmark to run
+--
 
 -- dobench :: Benchmark -> ParOrSeq -> Maybe Int -> IO ()
 dobench :: Benchmark -> ParOrSeq -> Maybe Int -> Int -> IO ()
@@ -102,13 +158,13 @@ dobench bench parorseq mb_size iters = do
         pure (A.size arr, A.size res0, tmed0, tall0)
       CSort alg -> do
         arr <- getInputAsList alg mb_size
-        (res0, tmed0, tall0) <- M.benchAndRunCSorts alg arr iters
+        (res0, tmed0, tall0) <- M.benchAndRunCSorts (sortFnC alg) arr iters
         unless (isSorted (res0)) (error $ show bench ++ ": result not sorted.")
         putStrLn "Sorted: OK"
         pure (length arr, length res0, tmed0, tall0)
       CxxSort alg -> do
         arr <- getInputAsList alg mb_size
-        (res0, tmed0, tall0) <- M.benchAndRunCxxSorts alg arr iters
+        (res0, tmed0, tall0) <- M.benchAndRunCxxSorts (sortFnCxx alg) arr iters
         unless (isSorted (res0)) (error $ show bench ++ ": result not sorted.")
         putStrLn "Sorted: OK"
         pure (length arr, length res0, tmed0, tall0)
